requirements.txt
spacy
scikit-learn
pdfplumber
python-docx
openai
python-dotenv
pandas